Move cannot fight message once again (onto CombatParticipant) ++ add some default BaseStats returns for no CharacterProperties to handle viewing in editor mode     -- since CombatParticipant tries to get stats off of it
