+++ Assets/Localization/Table_Core/Core_en.asset
@@ -238,7 +238,7 @@ MonoBehaviour:
     m_Localized: FRAIL
     m_Metadata:
       m_Items: []
-  - m_Id: 32539111364485120
+  - m_Id: 32974150610968576
     m_Localized: '{0} is wounded and cannot fight.'
     m_Metadata:
       m_Items: []
