+++ Assets/Localization/Table_Core/Core_fr.asset
@@ -238,7 +238,7 @@ MonoBehaviour:
     m_Localized: FRAGILE
     m_Metadata:
       m_Items: []
-  - m_Id: 32539111364485120
+  - m_Id: 32974150610968576
     m_Localized: "{0} est bless\xE9 et ne peut pas combattre."
     m_Metadata:
       m_Items: []
